Make Sauce metadata handling more robust
If errors occur when updating the Sauce Labs REST API, they're
silently swallowed.

Furthermore, if errors occur during parsing of `sauceEnd.js`, they
_also_ fail silently, leading users unsure of what's going on.

This change adds error handling which prints to `stderr` and
informs the user on `stdout` of a problem.

It also moves the Jenkins plugin log line above the REST API call,
so that this line is still printed if the API has issues.